// lib/api/generated/schemas/adminListDto.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend Admin API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminFilterDto } from './adminFilterDto';
import type { AdminPaginationDto } from './adminPaginationDto';
import type { AdminSortDto } from './adminSortDto';

export interface AdminListDto {
  filter?: AdminFilterDto;
  pagination?: AdminPaginationDto;
  sort?: AdminSortDto;
}